feat: Add Box contract for storing numbers
This commit adds a new contract called Box.sol, which allows for storing and retrieving numbers. The contract is based on the OpenZeppelin Ownable contract and includes a store function for setting the number and a getNumber function for retrieving it. This feature will be useful for future functionality in the Foundry Dao project.