refactor: Add mint function to GovernToken contract
This commit adds a new function called `mint` to the `GovernToken` contract in `GovernToken.sol`. The `mint` function allows for the creation of new tokens by minting them and assigning them to a specified address. This feature will be useful for various token distribution scenarios in the Foundry Dao project.